refactor(Menu): Move menu to separate component
